Refactor github sync with cron
